Moved determineFilePathOrDatasetFile to OperatorFilePathUtils.scala and made all inherited classes use the path util
