+++ core/amber/src/main/scala/edu/uci/ics/texera/workflow/operators/util/OperatorFilePathUtils.scala
@@ -1,0 +1,25 @@
+package edu.uci.ics.texera.workflow.operators.util
+
+import edu.uci.ics.amber.engine.common.AmberConfig
+import edu.uci.ics.amber.engine.common.storage.DatasetFileDocument
+
+import java.nio.file.Paths
+
+object OperatorFilePathUtils {
+  // resolve the file path based on whether the user system is enabled
+  // it will check for the presence of the given filePath/Desc based on fileName
+  def determineFilePathOrDatasetFile(fileName: Option[String]): (String, DatasetFileDocument) = {
+    if (AmberConfig.isUserSystemEnabled) {
+      // if user system is defined, a datasetFileDesc will be initialized, which is the handle of reading file from the dataset
+      val datasetFile = Some(new DatasetFileDocument(Paths.get(fileName.get)))
+      val file = datasetFile.getOrElse(
+        throw new RuntimeException("Dataset file descriptor is not provided.")
+      )
+      (null, file)
+    } else {
+      // otherwise, the fileName will be inputted by user, which is the filePath.
+      val filepath = fileName.getOrElse(throw new RuntimeException("File path is not provided."))
+      (filepath, null)
+    }
+  }
+}
